move lsl pull data timer, lsl visualization timer, create GUI component, workers, worker_thread to StreamWidget.py TODO: checkbox for plotting and group plots
